// packages/dashboard/src/http/types/GetMe.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export const getMeStatus200RoleEnum = {
    ADMIN: "ADMIN",
    MEMBER: "MEMBER"
} as const;

export type GetMeStatus200RoleEnumKey = (typeof getMeStatus200RoleEnum)[keyof typeof getMeStatus200RoleEnum];

export type GetMeStatus200 = {
    id: string;
    name: string;
    /**
     * @description
     * Format: `email`
     * @type string
    */
    email: string;
    role: GetMeStatus200RoleEnumKey;
    isActive: boolean;
};

export type GetMeOptions = {
    body?: never;
    path?: never;
    query?: never;
    headers?: never;
};

export type GetMeResponses = {
    "200": GetMeStatus200;
};

/**
 * @description Union of all possible responses
*/
export type GetMeResponse = GetMeStatus200;